Override `fetch` with array of override objects.
If you have multiple overrides that you're going to reuse, you can now
put them in an array. If you had a `Binder` to use as the base URL, plus
some special settings for connection pooling, you can place the `Binder`
in an array with an override object that has the connection pooling
settings.

Closes #31.