Implement `createWrappedDispatcher`.
The method chaining was silly. It came out of the notion that a service
would return the middle function, which was a way to expose the Connect
middleware without exposing Inlet. Subtle and unimportant. I have no
problem exposing Inlet. It's not like it is unmentionable or the mere
mention of Inlet makes these servers impossible to explain. It is not an
implementation detail, it is the interface. You're exposing it anyway
by extending the middleware function with a function.